Add files via upload
Create a Java class called Student having the fields regno  , cgpa and two other parameter as per your assumption. Enter all the details through Edit Text. Initialize and pass the object of Student class which store all the details on click of Save in Firebase Real-time Database and pass the control to Home Activity with Welcome Text (Now you need to add Something in the Welcome Text View According to your assumption like Welcome to the and Few button like Profile and Placement Portal. 
Following are the constraints.
1.If Student Click on Profile  pass the control to Student profile activity where student will set his/her image through Camera /gallery  and store the selected image from student to his/her firebase storage. In case  Student Click  on Back button take the control previous activity.[Make sure you select the image from gallery of your own as a student] 
2. If Student Click on Placement Portal pass the control  to Placement Activity named placement portal where student will read the detail of stored value on the click of Get Status from firebase database  like current cgpa and active status .If cgpa is more than 6.6 show the list of upcoming drives must be visible in the list with Company name and Logo  .